Moved api routes to /api
