+++ lib/middlewares/auth.js
@@ -13,4 +13,15 @@ function authorize() {
   }
 }
 
-module.exports = { authorize }
+function requireSession({ session }, res, next) {
+  if (session.account) {
+    next()
+  } else {
+    res.redirect('/login')
+  }
+}
+
+module.exports = {
+  authorize,
+  requireSession
+}
