+++ lib/routes/accounts.js
@@ -1,134 +1,99 @@
 const { Router } = require('express')
+const { create, login, setPdsProvider } = require('../services/accounts')
+const { camelCase } = require('changecase-objects')
 const createError = require('http-errors')
-const SSEStream = require('ssestream')
-
-const { create, get, del } = require('../services/accounts')
-const consentService = require('../services/consents')
-const dataService = require('../services/data')
-const { authorize } = require('../middlewares/auth')
+const dropbox = require('../adapters/pds/dropbox')
+const {requireSession} = require('../middlewares/auth')
 
 const router = Router()
 
-// Create account
-router.post('/', async ({ originalUrl, body: account }, res, next) => {
+// Login
+router.get('/login', ({ query, session }, res, next) => {
+  if (query.redirect_uri) {
+    session.requestingService = camelCase(query)
+  }
+  res.render('login', {})
+})
+router.post('/login', async ({ body: { username, password }, session }, res, next) => {
+  let account
   try {
-    const result = await create(account)
-    res.status(201).send({
-      data: result,
-      links: {
-        self: `${originalUrl}/${encodeURIComponent(result.id)}`
-      }
-    })
-  } catch (error) {
-    if (error.name === 'ValidationError') {
-      next(createError(400, error))
+    account = await login(username, password)
+  } catch (err) {
+    return res.status(401).render('login', { error: 'Wrong username or password' })
+  }
+  try {
+    session.account = account
+
+    if (!account.pdsProvider) {
+      res.redirect('/settings')
+    } else if (session.requestingService) {
+      res.redirect('/authorize')
     } else {
-      next(error)
+      res.redirect('/')
     }
+  } catch (err) {
+    next(createError(err))
   }
 })
 
-// Get account
-router.get('/:accountId', authorize(), async ({ originalUrl, params: { accountId } }, res, next) => {
+// Register
+router.get('/register', (req, res, next) => {
+  res.render('register', {})
+})
+router.post('/register', async ({ body, session }, res, next) => {
   try {
-    const account = await get(accountId)
-    if (!account) {
-      return next(createError(404))
+    const account = await create(body)
+    session.account = account
+    res.redirect('/settings')
+  } catch (err) {
+    next(err)
+  }
+})
+
+// Settings
+router.get('/settings', requireSession, (req, res, next) => {
+  const providers = [
+    dropbox.description
+  ]
+  res.render('settings', { providers })
+})
+router.get('/settings/dropbox/callback', requireSession, async ({ query: { code }, session }, res, next) => {
+  try {
+    const credentials = await dropbox.authorize(code)
+    session.account = await setPdsProvider(session.account, 'dropbox', credentials)
+    if (session.requestingService) {
+      res.redirect('/authorize')
+    } else {
+      res.redirect('/')
     }
-    res.send({
-      data: account,
-      links: {
-        self: originalUrl
-      }
-    })
-  } catch (error) {
-    next(error)
+  } catch (err) {
+    next(err)
   }
 })
 
-// Delete account
-router.delete('/:accountId', async ({ originalUrl, params: { accountId } }, res, next) => {
+// Authorize
+router.get('/authorize', requireSession, (req, res, next) => {
   try {
-    const account = await del(accountId)
-    res.send(204)
-  } catch (error) {
-    next(error)
+    const token = createToken({ account })
+    res.redirect(`${req.session.requestingService.redirectURI}?access_token=${token}`)
+    // res.render('authorize', {})
+  } catch (err) {
+    next(err)
   }
 })
-
-// Fetch to consents
-router.get('/:accountId/consents', async (req, res, next) => {
-  const consents = await consentService.getForAccount(req.params.accountId)
-
-  res.status(200).send(consents.map(c => ({
-    data: c,
-    links: { self: `/consents/${encodeURIComponent(c.id)}` }
-  })))
+router.post('/authorize', requireSession, (req, res, next) => {
+  //
 })
 
-// Subscribe to stuff
-router.get('/:accountId/events', (req, res, next) => {
-  try {
-    const sseStream = new SSEStream(req)
-    const consentListener = consentService.subscribe(req.params.accountId, (data) => {
-      sseStream.write({
-        event: 'consent',
-        data
-      })
-    })
-
-    sseStream.pipe(res)
-    res.on('close', () => {
-      sseStream.unpipe(res)
-      consentListener.unsubscribe()
-    })
-  } catch (error) {
-    next(error)
-  }
-})
-
-// Get your mydata
-router.get('/:accountId/data', async (req, res, next) => {
-  const data = await dataService.getForAccount(req.params.accountId)
-
-  if (!data) {
-    return next(createError(404))
-  }
-
-  res.status(200).send({
-    data,
-    links: { self: encodeURIComponent(`/${req.params.accountId}/data`) }
+// Logout
+router.get('/logout', ({ session }, res, next) => {
+  session.destroy(err => {
+    if (err) {
+      next(err)
+    } else {
+      res.redirect('/')
+    }
   })
-})
-
-// Put your mydata
-router.put('/:accountId/data/:area', async ({ body: data, params }, res, next) => {
-  dataService.updateAreaForAccount(params.accountId, params.area, data)
-
-  res.status(200).send({
-    data,
-    links: { self: encodeURIComponent(`/${params.accountId}/data/${params.area}`) }
-  })
-})
-
-// Get attached devices
-router.get('/:accountId/devices', (req, res, next) => {
-  res.send()
-})
-
-// Attach device
-router.post('/:accountId/devices', (req, res, next) => {
-  res.send()
-})
-
-// Get attached device
-router.get('/:accountId/devices/:deviceId', (req, res, next) => {
-  res.send()
-})
-
-// Detach device
-router.delete('/:accountId/devices/:deviceId', (req, res, next) => {
-  res.send()
 })
 
 module.exports = router
