+++ lib/routes/api/consents.js
@@ -1,6 +1,6 @@
 const { Router } = require('express')
 const createError = require('http-errors')
-const consentsService = require('../services/consents')
+const consentsService = require('../../services/consents')
 
 const router = Router()
