+++ lib/routes/api/index.js
@@ -1,15 +1,14 @@
 const { Router } = require('express')
-const api = require('./api')
 const accounts = require('./accounts')
+const consents = require('./consents')
 
 const router = Router()
-
-router.use('/', accounts)
-router.use('/api', api)
+router.use('/accounts', accounts)
+router.use('/consents', consents)
 
 /* GET home page. */
 router.get('/', (req, res, next) => {
-  res.render('index', { title: 'Express' })
+  res.send({})
 })
 
 module.exports = router
